+++ src/init.c
@@ -14,8 +14,8 @@ extern SEXP zstd_decompress_(SEXP src_, SEXP type_, SEXP dctx_, SEXP opts_, SEXP
 extern SEXP zstd_compress_stream_file_(SEXP robj, SEXP file_, SEXP cctx_, SEXP opts_);
 extern SEXP zstd_decompress_stream_file_(SEXP raw_vec_, SEXP type_, SEXP dctx_, SEXP opts_);
 
-extern SEXP zstd_serialize_(SEXP robj_, SEXP file_, SEXP cctx_, SEXP opts_);
-extern SEXP zstd_unserialize_(SEXP src_, SEXP dctx_, SEXP opts_);
+extern SEXP zstd_serialize_(SEXP robj_, SEXP file_, SEXP cctx_, SEXP opts_, SEXP use_file_streaming_);
+extern SEXP zstd_unserialize_(SEXP src_, SEXP dctx_, SEXP opts_, SEXP use_file_streaming_);
 
 extern SEXP zstd_serialize_stream_file_(SEXP robj, SEXP file_, SEXP cctx_, SEXP opts_);
 extern SEXP zstd_unserialize_stream_file_(SEXP raw_vec_, SEXP dctx_, SEXP opts_);
@@ -43,8 +43,8 @@ static const R_CallMethodDef CEntries[] = {
   {"zstd_compress_stream_file_"   , (DL_FUNC) &zstd_compress_stream_file_   , 4},
   {"zstd_decompress_stream_file_" , (DL_FUNC) &zstd_decompress_stream_file_ , 4},
   
-  {"zstd_serialize_"              , (DL_FUNC) &zstd_serialize_              , 4},
-  {"zstd_unserialize_"            , (DL_FUNC) &zstd_unserialize_            , 3},
+  {"zstd_serialize_"              , (DL_FUNC) &zstd_serialize_              , 5},
+  {"zstd_unserialize_"            , (DL_FUNC) &zstd_unserialize_            , 4},
   
   {"zstd_serialize_stream_file_"  , (DL_FUNC) &zstd_serialize_stream_file_  , 4},
   {"zstd_unserialize_stream_file_", (DL_FUNC) &zstd_unserialize_stream_file_, 3},