+++ NEWS.md
@@ -1,6 +1,11 @@
 
 
-# zstdlite 0.2.4.9007 2024-03-02
+# zstdlite 0.2.4.9009 2024-03-04
+
+* Streaming now optional for file I/O for `zstd_serialize()`/`zstd_unserialize()`
+    * Use option `use_file_streaming = TRUE` to enable
+
+# zstdlite 0.2.4.9008 2024-03-02
 
 * Added optional streaming for `zstd_compress()`/`zstd_decompress()`
     * Use option `use_file_streaming = TRUE` to enable